test(app-e2e): split npmx runtime fixtures
